로그아웃 기능 예외처리

+++ module-api/src/main/java/com/junior/service/member/OAuth2Service.java
@@ -9,11 +9,15 @@ import com.junior.dto.jwt.RefreshTokenDto;
 import com.junior.dto.member.CheckActiveMemberDto;
 import com.junior.dto.oauth2.OAuth2Provider;
 import com.junior.dto.oauth2.OAuth2UserInfo;
+import com.junior.exception.JwtErrorException;
+import com.junior.exception.StatusCode;
 import com.junior.repository.member.MemberRepository;
 import com.junior.security.JwtUtil;
 import com.junior.strategy.oauth2.KakaoOAuth2LoginStrategy;
 import com.junior.strategy.oauth2.OAuth2UserGenerator;
 import com.junior.util.RedisUtil;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.JwtException;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -51,7 +55,29 @@ public class OAuth2Service {
 
     public void logout(RefreshTokenDto refreshTokenDto) {
 
+
+        if (!refreshTokenDto.refreshToken().startsWith("Bearer ")) {
+            throw new JwtErrorException(StatusCode.INVALID_TOKEN);
+        }
+
         String refreshToken = refreshTokenDto.refreshToken().split(" ")[1];
+
+
+        try {
+            // 토큰이 access인지 확인 (발급시 페이로드에 명시)
+            String category = jwtUtil.getCategory(refreshToken);
+
+            if (!category.equals("refresh")) {
+                throw new JwtErrorException(StatusCode.NOT_REFRESH_TOKEN);
+            }
+        } catch (ExpiredJwtException e) {
+            // 토큰 만료 여부 확인
+            throw new JwtErrorException(StatusCode.EXPIRED_REFRESH_TOKEN);
+        } catch (JwtException | IllegalArgumentException e) {
+            //유효하지 않은 토큰
+            throw new JwtErrorException(StatusCode.INVALID_TOKEN);
+        }
+
 
         redisUtil.deleteData(refreshToken);
